requirements.txt
Flask==2.3.2
prometheus-client==0.17.0
python-json-logger==2.0.7
gunicorn==20.1.0
pytest==7.4.0
requests==2.31.0
